requirements.txt
streamlit
opencv-python
numpy
Pillow
torch
torchvision
timm
scikit-learn 
matplotlib
joblib 